+++ Assets/Scripts/Magic/ActualMagic/Lance.cs
@@ -8,7 +8,7 @@ namespace HMF.Thesis.Magic.ActualMagic
     {
         public override string Name => "Lance";
 
-        public override void Use(string[] tagsToIgnore, MagicFocusData magicFocus, Vector2 center, int dir)
+        public override void Use(string[] tagsToIgnore, MagicFocusData magicFocus, Vector2 center, float dir)
         {
             var point = new Vector2(center.x + dir * magicFocus.attackRange / 2, center.y);
             var colliders = Physics2D.OverlapAreaAll(center, new Vector2(center.x + magicFocus.attackRange, magicFocus.attackRange / 2));
